# openapi.yaml
openapi: 3.0.3
info:
  title: Expense Tracker API
  version: 1.4.1
  description: REST API for expense tracking with PostgreSQL, JWT auth, role-based
    access, and request correlation via x-request-id. If the client sends x-request-id
    the API reuses it; otherwise the server generates one. Each request is logged
    on response finish with requestId, method, path, statusCode and durationMs.
servers:
- url: http://localhost:3000
  description: Local (Node) / Docker Compose
tags:
- name: Health
- name: Auth
- name: Users
- name: Projects
- name: Expenses
paths:
  /api/v1/health:
    get:
      summary: Health check
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      status:
                        type: string
                        example: ok
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
      tags:
      - Health
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
  /api/v1/auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRegisterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthRegisterResponse'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '409':
          description: Email already in use
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
      tags:
      - Auth
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
  /api/v1/auth/login:
    post:
      summary: Login (returns JWT)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthLoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthLoginResponse'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
      tags:
      - Auth
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
  /api/v1/users/me:
    get:
      summary: Get current user
      security:
      - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserMeResponse'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '401':
          description: Unauthorized (missing/invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
      tags:
      - Users
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
  /api/v1/projects:
    get:
      summary: List projects
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
      - in: query
        name: page
        schema:
          type: integer
          minimum: 1
          default: 1
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 20
          default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectsListResponse'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
      tags:
      - Projects
    post:
      summary: Create project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
      tags:
      - Projects
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
  /api/v1/projects/{id}:
    get:
      summary: Get project by id
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
      tags:
      - Projects
    put:
      summary: Update project by id
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
      tags:
      - Projects
    delete:
      summary: Delete project by id
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '204':
          description: No content
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '404':
          description: Not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
      tags:
      - Projects
  /api/v1/users:
    get:
      summary: List users (admin only)
      tags:
      - Users
      security:
      - bearerAuth: []
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsersListResponse'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '401':
          description: Unauthorized (missing/invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '403':
          description: Forbidden (insufficient role)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
  /api/v1/expenses:
    post:
      summary: Create an expense
      tags:
      - Expenses
      security:
      - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpenseCreate'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '401':
          description: Unauthorized (missing/invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
    get:
      summary: List expenses (paginated)
      tags:
      - Expenses
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
      - in: query
        name: page
        schema:
          type: integer
          minimum: 1
          default: 1
      - in: query
        name: limit
        schema:
          type: integer
          minimum: 1
          maximum: 20
          default: 10
      - in: query
        name: from
        description: Start of occurred_at range. Accepts ISO 8601 date (`YYYY-MM-DD`)
          or datetime.
        schema:
          type: string
          example: '2024-02-01'
      - in: query
        name: to
        description: End of occurred_at range. Accepts ISO 8601 date (`YYYY-MM-DD`)
          or datetime.
        schema:
          type: string
          example: '2024-02-29'
      - in: query
        name: order
        description: Sort direction applied to occurred_at and id.
        schema:
          type: string
          enum:
          - asc
          - desc
          default: desc
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpensesListResponse'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '401':
          description: Unauthorized (missing/invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
  /api/v1/expenses/{id}:
    get:
      summary: Get an expense by id (owner-only)
      tags:
      - Expenses
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseReadResponse'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '401':
          description: Unauthorized (missing/invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '404':
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
    patch:
      summary: Update an expense (owner-only, partial)
      tags:
      - Expenses
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExpensePatch'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExpenseResponse'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '401':
          description: Unauthorized (missing/invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '404':
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
    delete:
      summary: Delete an expense (owner-only)
      tags:
      - Expenses
      security:
      - bearerAuth: []
      parameters:
      - $ref: '#/components/parameters/XRequestIdRequestHeader'
      - in: path
        name: id
        required: true
        schema:
          type: integer
          minimum: 1
      responses:
        '204':
          description: No Content
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '401':
          description: Unauthorized (missing/invalid token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
        '404':
          description: Expense not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
          headers:
            x-request-id:
              $ref: '#/components/headers/XRequestId'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ApiError:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Invalid request body
            requestId:
              type: string
              description: Request correlation id. Same value as response header x-request-id.
              example: d304bb93-3c35-4bd5-a6a1-c422431f0bd9
      required:
      - error
    Project:
      type: object
      properties:
        id:
          type: integer
          example: 1
        name:
          type: string
          example: Project A
        price_cents:
          type: integer
          example: 1000
        created_at:
          type: string
          format: date-time
          example: '2026-02-09T17:04:05.154Z'
      required:
      - id
      - name
      - price_cents
      - created_at
    ProjectCreate:
      type: object
      properties:
        name:
          type: string
          example: Project X
        price_cents:
          type: integer
          example: 1234
      required:
      - name
      - price_cents
    ProjectsListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Project'
        meta:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            total:
              type: integer
              example: 5
      required:
      - data
      - meta
    ProjectResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Project'
      required:
      - data
    AuthRegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: user@mail.com
        password:
          type: string
          minLength: 8
          example: '12345678'
      required:
      - email
      - password
    AuthLoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: user@mail.com
        password:
          type: string
          minLength: 8
          example: '12345678'
      required:
      - email
      - password
    UserRegistered:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@mail.com
        role:
          type: string
          example: user
        created_at:
          type: string
          format: date-time
          example: '2026-02-18T10:00:00.000Z'
      required:
      - id
      - email
      - role
      - created_at
    AuthRegisterResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UserRegistered'
      required:
      - data
    AuthUserInToken:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@mail.com
      required:
      - id
      - email
    AuthLoginResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            token:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            user:
              $ref: '#/components/schemas/AuthUserInToken'
          required:
          - token
          - user
      required:
      - data
    UserMe:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: user@mail.com
        created_at:
          type: string
          format: date-time
          example: '2026-02-18T10:00:00.000Z'
      required:
      - id
      - email
      - created_at
    UserMeResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/UserMe'
      required:
      - data
    UserAdmin:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        email:
          type: string
          example: admin@mail.com
        role:
          type: string
          example: admin
        created_at:
          type: string
          format: date-time
          example: '2026-02-23T10:00:00.000Z'
      required:
      - id
      - email
      - role
      - created_at
    UsersListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/UserAdmin'
      required:
      - data
    Expense:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        user_id:
          type: integer
          format: int64
          example: 1
        amount_cents:
          type: integer
          example: 1234
          minimum: 1
        description:
          type: string
          example: Coffee
        occurred_at:
          type: string
          format: date-time
          example: '2026-02-23T10:00:00.000Z'
        created_at:
          type: string
          format: date-time
          example: '2026-02-23T10:00:00.000Z'
      required:
      - id
      - user_id
      - amount_cents
      - description
      - occurred_at
      - created_at
    ExpenseRead:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 10
        amount_cents:
          type: integer
          example: 1234
          minimum: 1
        description:
          type: string
          example: Coffee
        occurred_at:
          type: string
          format: date-time
          example: '2026-02-23T10:00:00.000Z'
        created_at:
          type: string
          format: date-time
          example: '2026-02-23T10:00:00.000Z'
      required:
      - id
      - amount_cents
      - description
      - occurred_at
      - created_at
    ExpenseReadResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ExpenseRead'
      required:
      - data
    ExpenseCreate:
      type: object
      properties:
        amount_cents:
          type: integer
          minimum: 1
          example: 1234
        description:
          type: string
          minLength: 1
          maxLength: 200
          example: Coffee
        occurred_at:
          type: string
          format: date-time
          example: '2026-02-23T10:00:00.000Z'
      required:
      - amount_cents
      - description
    ExpensePatch:
      type: object
      minProperties: 1
      additionalProperties: false
      properties:
        amount_cents:
          type: integer
          minimum: 1
          example: 1500
        description:
          type: string
          minLength: 1
          maxLength: 200
          example: Coffee beans
        occurred_at:
          type: string
          format: date-time
          example: '2026-02-23T10:00:00.000Z'
    ExpenseResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/Expense'
      required:
      - data
    ExpensesListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Expense'
        meta:
          type: object
          properties:
            page:
              type: integer
              example: 1
            limit:
              type: integer
              example: 10
            total:
              type: integer
              example: 5
      required:
      - data
      - meta
  headers:
    XRequestId:
      description: Final request correlation identifier. Matches error.requestId in
        error responses and the requestId field written to server logs.
      schema:
        type: string
  parameters:
    XRequestIdRequestHeader:
      in: header
      name: x-request-id
      required: false
      description: Optional client-supplied request correlation id. If omitted, the
        server generates one. The final value is returned in response header x-request-id
        and used in server logs.
      schema:
        type: string